pyproject.toml
[project]
name = "demo-fastapi"
version = "0.1.0"
description = "A FastAPI project with Strawberry GraphQL integration"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "strawberry-graphql==0.219.1",
    "pydantic==2.6.1",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.3.0",
    "mypy==1.8.0",
]


[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true 